@@ -1,6 +1,7 @@
 package kommando_test
 
 import (
+	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -8,8 +9,36 @@ import (
 	"testing"
 )
 
+func TestMain(m *testing.M) {
+	// Build binary
+	cwd, err := os.Getwd()
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Failed to get current directory: %v\n", err)
+		os.Exit(1)
+	}
+
+	exampleDir := filepath.Join(cwd, "example")
+	binaryPath := filepath.Join(exampleDir, "kommando")
+
+	// Build the kommando binary
+	cmd := exec.Command("go", "build", "-o", binaryPath, "./cmd/kommando/...")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		fmt.Fprintf(os.Stderr, "Failed to build kommando binary: %v\n", err)
+		os.Exit(1)
+	}
+
+	// Run tests
+	code := m.Run()
+
+	// Cleanup
+	os.Remove(binaryPath)
+
+	os.Exit(code)
+}
+
 func TestKommandoCLI(t *testing.T) {
-	// Binary is built and copied by Makefile build target before tests run
 	cwd, err := os.Getwd()
 	if err != nil {
 		t.Fatalf("Failed to get current directory: %v", err)
